% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBetaCon.R
\name{myBetaCon}
\alias{myBetaCon}
\title{myBetaCon}
\usage{
myBetaCon(a, b, n, x)
}
\arguments{
\item{a}{shape 1}

\item{b}{shape 2}

\item{n}{number of trials}

\item{x}{number of successes}
}
\value{
a plot of the posterior distribution
}
\description{
Produces a plot of the posterior distribution for the Beta-Binomial conjugate pair
}
\examples{
\dontrun{myBetaCon(a=5, b=7, n=10, x=4)}
}
