% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGamCon.R
\name{myGamCon}
\alias{myGamCon}
\title{myGamCon}
\usage{
myGamCon(a, b, n, x)
}
\arguments{
\item{a}{shape 1}

\item{b}{shape 2}

\item{n}{number of trials}

\item{x}{number of successes}
}
\value{
a plot of the posterior distribution
}
\description{
A plot of the posterior of a Gamma-Poisson distribution.
}
\examples{
\dontrun{myGamCon(a=5, b=7, n=10, x=4)}
}
