% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybeta.R
\name{shinybeta}
\alias{shinybeta}
\title{shinybeta}
\usage{
shinybeta()
}
\value{
an interactive plot
}
\description{
An interactive app which produces a plot of the posterior
}
\examples{
\dontrun{shinycoins()}
}
